requirements.txt
Django==5.1.2
django-filter==24.3
django_resized==1.0.2
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
django-import-export==4.1.1
django-redis==5.4.0
django-leaflet==0.30.1
django-reversion==5.1.0
redis==5.1.1
flake8==7.1.1
psycopg2-binary==2.9.9
Pillow==10.4.0
drf-yasg==1.21.7
Faker==30.3.0
django-debug-toolbar==4.4.6
Khayyam==3.0.17
uwsgi==2.0.27
setuptools==75.1.0
requests==2.32.3